the fianl version (1st) of the game
have removed the 1st algorithm for win checking (mine) and the ai's
and then used the idea of the one from stackoverflow and the checkWin
funciton now works greatly and it is good.
changed the HTML and CSS to make a current player indication above the
game board (this is what changed in the HTML, CSS).
the game is now working correctly and it is great.

things you might want to do
---------------

- [ ] make it a responsive design (make it good-looking on mobiles)
- [ ] use service workers to make it work offline
- [ ] try to make an animaiton of dropping the piece (cell) into it's
      place not just appearing where it should be placed
- [ ] host it on github (github pages)
[Ticket: 01-00]

Committer: Codes_Broker <andrew@pop-os.localdomain>

On branch main
Changes to be committed:
      modified:   css/master.css
      modified:   index.html
      modified:   js/master.js